+++ src/Symfony/Component/Form/Extension/Core/DataMapper/RadioListMapper.php
@@ -28,7 +28,7 @@ class RadioListMapper implements DataMapperInterface
     /**
      * {@inheritdoc}
      */
-    public function mapDataToForms($choice, iterable $radios)
+    public function mapDataToForms($choice, \Traversable $radios)
     {
         if (\is_array($radios)) {
             trigger_deprecation('symfony/form', '5.3', 'Passing an array as the second argument of the "%s()" method is deprecated, pass "\Traversable" instead.', __METHOD__);
@@ -47,12 +47,8 @@ class RadioListMapper implements DataMapperInterface
     /**
      * {@inheritdoc}
      */
-    public function mapFormsToData(iterable $radios, &$choice)
+    public function mapFormsToData(\Traversable $radios, &$choice)
     {
-        if (\is_array($radios)) {
-            trigger_deprecation('symfony/form', '5.3', 'Passing an array as the first argument of the "%s()" method is deprecated, pass "\Traversable" instead.', __METHOD__);
-        }
-
         if (null !== $choice && !\is_string($choice)) {
             throw new UnexpectedTypeException($choice, 'null or string');
         }