+++ src/Symfony/Component/Form/FormConfigInterface.php
@@ -18,8 +18,6 @@ use Symfony\Component\PropertyAccess\PropertyPathInterface;
  * The configuration of a {@link Form} object.
  *
  * @author Bernhard Schussek <bschussek@gmail.com>
- *
- * @method callable|null getIsEmptyCallback() Returns a callable that takes the model data as argument and that returns if it is empty or not - not implementing it is deprecated since Symfony 5.1
  */
 interface FormConfigInterface
 {
@@ -247,4 +245,9 @@ interface FormConfigInterface
      * @return mixed The option value
      */
     public function getOption(string $name, $default = null);
+
+    /**
+     * Returns a callable that takes the model data as argument and that returns if it is empty or not.
+     */
+    public function getIsEmptyCallback(): ?callable;
 }